requirements.txt
torch==2.6.0
transformers==4.51.3
pillow==11.1.0
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.17
qdrant-client==1.14.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
pyodbc==5.2.0
pymssql
requests==2.32.4
apscheduler==3.11.0
pillow-avif-plugin==1.5.2
# spacy==3.7.2

